Renames AposTableMixin to AposManagerMixin
