+++ modules/@apostrophecms/modal/ui/apos/mixins/AposManagerMixin.js
